Add database seeding during app startup
Introduced a call to `SeedService.SeedDatabase(app.Services)`
to perform database seeding during the application's startup
process. This ensures the database is pre-populated with
necessary initial data. Added the `await` keyword to make the
seeding operation asynchronous. Included a `using` directive
for the `UserRoles.Services` namespace to enable the use of
the `SeedService` class.